--- Muzyk-API/DTOS/UserForLoginDto.cs
@@ -1,8 +1,0 @@
-namespace Muzyk_API.DTOS
-{
-    public class UserForLoginDto
-    {
-        public string Username { get; set; }    
-        public string Password { get; set; }
-    }
-}
