--- Muzyk-API/DTOS/UserForUpdateDto.cs
@@ -1,12 +1,0 @@
-namespace Muzyk_API.DTOS
-{
-    public class UserForUpdateDto
-    {
-        public string Introduction { get; set; }
-        public string Interests { get; set; }
-        public string City { get; set; }
-        public string Country { get; set; }
-        public string Genre { get; set; }
-        public int YearsOfExperience { get; set; }
-    }
-}
