+++ Muzyk-DTOS/DTOS/MessageForCreationDto.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace Muzyk_DTOS
+{
+    public class MessageForCreationDto
+    {
+        public int SenderId { get; set; }
+        public int RecepientId { get; set; }
+        public DateTime MessageSent { get; set; }
+        public string Content { get; set; }
+        public MessageForCreationDto()
+        {
+            MessageSent = DateTime.Now;
+        }
+    }
+}
