+++ Muzyk-DTOS/DTOS/UserForLoginDto.cs
@@ -1,0 +1,8 @@
+namespace Muzyk_DTOS
+{
+    public class UserForLoginDto
+    {
+        public string Username { get; set; }    
+        public string Password { get; set; }
+    }
+}
