+++ Muzyk-DTOS/DTOS/VideoForDetailDto.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Muzyk_DTOS
+{
+    public class VideoForDetailDto
+    {
+        public int Id { get; set; }
+        public string VideoUrl { get; set; }
+        public string Description { get; set; }
+        public DateTime DateAdded { get; set; }
+    }
+}
